Prevent circular import error
